backend/requirements.txt
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
Django==6.0.3
djangorestframework==3.17.0
flet==0.82.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
msgpack==1.1.2
oauthlib==3.3.1
pillow==12.1.1
psycopg2-binary==2.9.11
python-decouple==3.8
repath==0.9.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
